+++ bin/coal.dart
@@ -1,0 +1,7 @@
+import 'dart:io';
+
+import 'package:coal/src/cli/cli.dart';
+
+Future<void> main(List<String> args) async {
+  exitCode = await runCoal(args);
+}